modificação pacote serviço peças

+++ src/main/java/com/pecassystem/pecas/servico/PecaServico.java
@@ -1,11 +1,10 @@
 package com.pecassystem.pecas.servico;
 
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
+
 import org.springframework.stereotype.Service;
 import com.pecassystem.pecas.modelo.Peca;
-import com.pecassystem.pecas.modelo.RespostaModelo;
+
 import com.pecassystem.pecas.repositorio.PecaRepositorio;
 
 @Service
@@ -14,8 +13,7 @@ public class PecaServico {
     @Autowired
     private PecaRepositorio pecaRepositorio;
 
-    @Autowired
-    private RespostaModelo respostaModelo;
+    
 
     //Lista todas as peças
     public Iterable<Peca> listar(){
@@ -24,37 +22,32 @@ public class PecaServico {
 
     
 
-        public ResponseEntity<?> cadastrar(Peca peca) {
-            try {
-                // Salva a peça no banco de dados
-                Peca novaPeca = pecaRepositorio.save(peca);
-                return new ResponseEntity<>(novaPeca, HttpStatus.CREATED);
-            } catch (Exception e) {
-                // Tratamento de erros de persistência ou outros erros
-                respostaModelo.setMensagen("Erro ao salvar a peça: " + e.getMessage());
-                return new ResponseEntity<>(respostaModelo, HttpStatus.INTERNAL_SERVER_ERROR);
-            }
+    public Peca cadastrar(Peca peca) {
+        try {
+            return pecaRepositorio.save(peca);
+        } catch (Exception e) {
+            throw new RuntimeException("Erro ao salvar a peça: " + e.getMessage());
         }
+    }
+    
 
-        public ResponseEntity<?> alterar(Peca peca) {
-            try {
-                // Salva a peça no banco de dados
-                Peca novaPeca = pecaRepositorio.save(peca);
-                return new ResponseEntity<>(novaPeca, HttpStatus.OK);
-            } catch (Exception e) {
-                // Tratamento de erros de persistência ou outros erros
-                respostaModelo.setMensagen("Erro ao salvar a peça: " + e.getMessage());
-                return new ResponseEntity<>(respostaModelo, HttpStatus.INTERNAL_SERVER_ERROR);
-            }
+    public Peca alterar(Peca peca) {
+        try {
+            return pecaRepositorio.save(peca);
+        } catch (Exception e) {
+            throw new RuntimeException("Erro ao alterar a peça: " + e.getMessage());
         }
+    }
+    
 
-        public ResponseEntity<RespostaModelo> remover(int id){
-
+    public void remover(int id) {
+        if (pecaRepositorio.existsById(id)) {
             pecaRepositorio.deleteById(id);
-
-            respostaModelo.setMensagen("Peça removida com Sucesso");
-            return new ResponseEntity<RespostaModelo>(respostaModelo,HttpStatus.OK);
+        } else {
+            throw new RuntimeException("Peça não encontrada com o ID: " + id);
         }
+    }
+
